git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/tube-cpa@48039 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cfa/simplification/ExpressionSimplificationVisitor.java
@@ -291,7 +291,7 @@ public class ExpressionSimplificationVisitor
             }
             return new CFloatLiteralExpression(loc, exprType, BigDecimal.valueOf(v));
           default:
-            // fall-through and return the original expression
+            // return the original expression below
         }
 
       } else if (unaryOperator == UnaryOperator.TILDE
@@ -337,8 +337,8 @@ public class ExpressionSimplificationVisitor
 
     // enum constant
     if (decl instanceof CEnumerator) {
-      final long v = ((CEnumerator) decl).getValue();
-      return new CIntegerLiteralExpression(expr.getFileLocation(), type, BigInteger.valueOf(v));
+      final BigInteger v = ((CEnumerator) decl).getValue();
+      return new CIntegerLiteralExpression(expr.getFileLocation(), type, v);
     }
 
     // const variable, inline initializer
@@ -359,7 +359,7 @@ public class ExpressionSimplificationVisitor
             case DOUBLE:
               return new CFloatLiteralExpression(expr.getFileLocation(), type, v.bigDecimalValue());
             default:
-              // fall-through and return the original expression
+              // return the original expression below
           }
         }
       }
